Migrate offset from Long to Int : integrate changes from last libcore version

+++ src/main/scala/co/ledger/wallet/daemon/controllers/AccountsController.scala
@@ -204,7 +204,7 @@ class AccountsController @Inject()(accountsService: AccountsService) extends Con
 
 object AccountsController {
   private val DEFAULT_BATCH: Int = 20
-  private val DEFAULT_OFFSET: Long = 0
+  private val DEFAULT_OFFSET: Int = 0
   private val DEFAULT_OPERATION_MODE: Int = 0
 
 
@@ -336,7 +336,7 @@ object AccountsController {
                                   @RouteParam wallet_name: String,
                                   @RouteParam account_index: Int,
                                   @RouteParam token_address: String,
-                                  @QueryParam offset: Long = DEFAULT_OFFSET,
+                                  @QueryParam offset: Int = DEFAULT_OFFSET,
                                   @QueryParam batch: Int = DEFAULT_BATCH,
                                   request: Request
                                 )
@@ -375,7 +375,7 @@ object AccountsController {
                                 @RouteParam override val account_index: Int,
                                 @QueryParam next: Option[UUID],
                                 @QueryParam previous: Option[UUID],
-                                @QueryParam offset: Long = DEFAULT_OFFSET,
+                                @QueryParam offset: Int = DEFAULT_OFFSET,
                                 @QueryParam batch: Int = DEFAULT_BATCH,
                                 @QueryParam full_op: Int = DEFAULT_OPERATION_MODE,
                                 @QueryParam contract: Option[String],
